Added encoder to compile pipeline

+++ encoder/arducam_encoder.hpp
@@ -66,7 +66,7 @@ private:
 	std::mutex encode_mutex_;
 	std::condition_variable encode_cond_var_;
 	std::thread encode_thread_[NUM_ENC_THREADS];
-	void encodeArducam(EncodeItem &item, uint8_t *&encoded_buffer, size_t &buffer_len);
+	void encodeArducam(EncodeItem &item, uint16_t *&encoded_buffer, size_t &buffer_len);
 
 	struct OutputItem
 	{
